requirements.txt
langchain
langchain-groq
langchain-community
langchain-tavily
langchain-huggingface
langchain-text-splitters
langgraph
pymongo
faiss-cpu
sentence-transformers
python-dotenv
certifi
pypdf